Fixed a problem with the readOutputXMLFile method that was causing the outputToXML tests to fail.
The problem was the method was appending only '\n' as a newline character when reading in the XML file, which causes the test to fail on Windows machines, while working on OS X, since Windows uses carriage return AND line feed.

The solution is to append the String returned by System.getProperty("line.seperator") instead, which should return the current platform's newline character.