Bugfix: When creating a session from a corrupted project file, the context.createSession() method was throwing an exception (not returning the session), but the new session itself was still hanging around in the context's list of sessions.
When it came time to exit, the cleanup code for the exit action
was trying to close this zombie session and failing with an NPE
becuase the frame hadn't been created.  We've fixed this by making
sure the zombie session doesn't last when the load fails.

There are also a number of other issues that got fixed as a side
effect of this fix:

-The SessionLifecycleEvent API now works like a regular Java event
 API (it has an event object, which gets passed to listeners)
 
-Finished the old ArchitectFrame refactoring.. a bunch of session
 lifecycle methods were still stranded in ArchitectFrame. We moved
 them into the session.
 
-Got rid of the special OS X case for exiting the app (now all
 platforms exit in the same way)
 
-Made the System.exit() in the session context optional (it defaults
 to no), which will help with those who want to embed the Architect's
 API