Fixed the build file so that it will check for the SQL Power Library before compiling. If it cannot find the library, it will point the developer to the DEVELOPERS file on how to set it up. Otherwise, it will then compile the library and use it in the classpath when compiling the Architect and its tests.
(I would like to get ant to print out directions for for getting the library, but unfortunately, the failure message gets printed all on one line.)

The build.properties.example contains an example for the sqlpower.library.home property which points to the sqlpower library project. 

The DEVELOPERS doc now uses the 'getting the source' directions from the Wiki, and also how to get the sqlpower-library source.

In the future, it would be nice for the sqlpower library dependency to be handled automatically, but I don't think there's a nice clean way for Ant to do this, without making the build file even more confusing than it already is. It would be worth checking out if Apache Maven can help us out here.