Reworked and simplified the QFA facility.
The whole shebang bolied down to one generic class which was moved into the SQL Power Library as
ExceptionReport, plus a custom per-application exception handler.  We moved the Architect's custom exception handler into the swingui package, since it was going to be the only thing left in the old qfa package.