Solved an old bug in the PokeDBThread implementation: When it failed for any reason, it would add a SQLExceptionNode to the parent that failed to populate, but because this was not done on the Swing Event Dispatch Thread, the SQLObject childrenInserted event was never refired to the DBTree.
The reason we never noticed this problem before was that most PokeDB failures are instant, and happen before the user attempts to expand the tree node. This means there was no need for the treeNodesInserted event. However, when the exception is due to a network socket timeout, there is plenty of time to try and expand the tree node. In that case, the Swing thread often unblocks and notices the database node is empty just before the PokeDBThread inserts the SQLException node. Then the JTree layout cache remembers the node is empty, and the absence of the event means the user never sees any error message at all.

Solution: Reworked the PokeDBThread into a SPSwingWorker, and now the SQLExceptionNode is added in cleanup(), which is on the Swing EDT, and the tree gets notified as it should. Part of this solution caused us to move the pokeDatabase() method into DBTree (it was in ArchitectUtils), but nothing else was using that method anyway.