Initial commit demonstrating the use of ResourceBundle to externalize the message Strings in the ArchitectFrame class. This helps pave the way for future internationalization of the Power*Architect, allowing volunteers to submit translated messages for different locales.
I used Eclipse's 'Externalize Strings' feature to do this.

Here's a quick rundown of how it works:
All messages are stored in a messages_<language>_<locale>.properties file. You can also have generic message files that for just language, or just a generic messages file (like the message.properties file in this commit), and all that specify a locale and use the same language will actually inherit the messages from the one with just the language. So I could create a messages_en_CA.properties file that contains only the Strings that are particular to Canadian English, and it would inherit the rest from the generic English file.

Just to note, I have the initial messages.properties file in English because if a user is using a locale for which no message file exists, the application will not run. So it's not enough for to simply have a messages_en.properties for now. Effectively, this sets English as the default language for the application.

The Messages class in the package loads the properties file as a ResourceBundle, and then the classes in that package use the Message class to get the message Strings.

The neat thing is that Java will use messages for a user's specific language and locale settings (as set in the OS) if the message file for it exists.