A new business model for our cube generator! Our model is a slightly altered version of Mondrian's.  The differences are: - Instead of using arrays, we use GlazedLists so we can fire events. - All instance variables are now private and have getters and setters.
The MetaGenerator.java generated the MondrianDef based on the Mondrian.xml.  Currently, after the MondrianDef is generated, it is required to change the following lists (along with their EventLists and ObservableElementLists) to protected rather than private or it will not compile: 
ignoreColumns, foreignKeys, measures, levels, expressions  
It is also required to remove the all the occurrences of the statements "_parser = _parser;" and "_tempArray = _tempArray;".