Further decoupling of the PlayPen from SQLObjects and the DBTree. This time, we generalized column selection in TableTane to item selection in ContainerPane. Now all selection synchronization is handled by the synchronizer in RelationalPlayPenFactory.
This required us to implement an event system so the PlayPen (actually the PlayPenContentPane) can notify when components are added and removed. This new event type is called PlayPenContentEvent.

We also fixed a regression in column selection handling: When you select a single column and delete it, the next column should get selected instead of the selection reverting back to the whole table. And if you undo the deletion, the undeleted column should come back selected.