Changed OLAP Edit Sessions to live in a frame instead of a dialog. This makes it easier to use and less bizarre (for example, you can now get the architect frame in front of an olap editor).
One problem this has introduced is that the menu bar is attached to the Architect Frame. On Windows, this isn't a big deal: the OLAP sessions simply don't have a menu bar. But on OS X, it sucks because the screen menu bar goes blank when you're using an OLAP frame. We'll come up with a good way of dealing with this soon.