Fix for bug 1618. There is now a listener for adding and removing data sources to the tree in Architect. Data source adds and removes can now be undone and redone.
The listener added to the root of the data sources does not listen
to changes in the children of the data sources as the children of 
a data source are created by the populate method. Children should be lazily loaded and undoing a populate does not make sense. This may need to change when refresh is implemented for data sources.