This is a fix to the forum bug post about relationship line color on Windows machine. The problem was our Color Set doesn't include the default foreground color on Windows, so the color selector in RelationshipEditPanel picks up its first item, which is Red. Now add the color to our Color Set if it's not included.
The problem I haven't solved yet is this won't help if the file is created and saved on Windows and then use it on Mac.(the reason is we add color to our Color Set in the constructor, so getForegroundColor() method returns the general foreground instead of saved Windows Relationship foreground color) We can solve this problem by doing the same thing(add the color to our Color Set if it's not included) in the constructor of RelationshipEditPanel and createPop() of Relationship, but I don't think it's ideal. Any ideas?