NEW - bug 2722: Unable to open existing project http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2722
Added a prompt for users opening a file that was saved with a newer version of Architect. The user will be asked if they 
want to upgrade as well. To make the tests pass successfully the user prompter is being set to a default prompter
otherwise the tests will pause for user input.