NEW - bug 2721: Can't expand any database nodes in Copy Table Data dialogue http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2721
The DBTreeModel is populating columns when accessing children instead of the DBTree doing it for the model. 
To prevent the DBTreeModel from going into an infinite loop between populating a SQLTable and updating itself
from having children added the populate columns and indices on the SQLTable will exit early if it has already
been entered earlier on the same thread.