Adding in a stop-gap solution for moving columns in a table with snapshots.
The problem is moving a column causes a remove and an add. The remove will cause the listener to update the use count of the snapshot to 0 in some cases and remove the snapshot and UDT. When the column is added back in the UDT upstream of the column proxy UDT the old snapshot is snapshotted and a listener is attempted to be added to it as though it is a system snapshot and causes an NPE.

Now the use count is only updated at the end of a transaction which works for now because all moves happen in a transaction. A better solution would involve merging the snapshot object with the snapshotted UDT.